// src/api-chain/model/chainLocation.ts
export interface ChainLocation { 
    /**
     * address
     */
    address?: string;
    /**
     * city
     */
    city?: string;
    country?: ChainCountry;
    /**
     * state / province / region
     */
    state?: string;
    /**
     * ZIP / postal code / p.p. box
     */
    zip?: string;
    /**
     * location latitude
     */
    latitude?: number;
    /**
     * location longitude
     */
    longitude?: number;
    /**
     * Custom site coordinate
     */
    site?: string;
    /**
     * Custom sector coordinate
     */
    sector?: string;
    /**
     * Custom cell coordinate
     */
    cell?: string;
    /**
     * Custom village coordinate
     */
    village?: string;
    /**
     * Is location publicly visible?
     */
    isPubliclyVisible?: boolean;
}