// src/api-chain/model/chainSemiProductPrice.ts
/*
 * Copyright(c) 2009 - 2019 Abelium d.o.o.
 * Kajuhova 90, 1000 Ljubljana, Slovenia
 * All rights reserved
 * Copyright(c) 1995 - 2018 T-Systems Multimedia Solutions GmbH
 * Riesaer Str. 5, 01129 Dresden
 * All rights reserved.
 *
 * coffee-be
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the openapi-typescript-angular-generator.
 * https://github.com/alenabelium/openapi-typescript-angular-generator
 * Do not edit the class manually.
 */





export interface ChainSemiProductPrice { 
    semiProductId: string;
    price: number;
    currency: string;
}

/**
 * Namespace for property- and property-value-enumerations of ChainSemiProductPrice.
 */
export namespace ChainSemiProductPrice {
    /**
     * All properties of ChainSemiProductPrice.
     */
    export enum Properties {
        semiProductId = 'semiProductId',
        price = 'price',
        currency = 'currency'
    }


    export function formMetadata() {
        return  {
            metadata: formMetadata,
            classname: 'ChainSemiProductPrice',
            vars: [
                {
                    isReadOnly: false,
                    isEnum: false,
                    required: true,
                    name: 'semiProductId',
                    classname: 'ChainSemiProductPrice',
                    dataType: 'string',
                    isPrimitiveType: true,
                    isListContainer: false,
                    complexType: ''
                },
                {
                    isReadOnly: false,
                    isEnum: false,
                    required: true,
                    name: 'price',
                    classname: 'ChainSemiProductPrice',
                    dataType: 'number',
                    isPrimitiveType: true,
                    isListContainer: false,
                    complexType: ''
                },
                {
                    isReadOnly: false,
                    isEnum: false,
                    required: true,
                    name: 'currency',
                    classname: 'ChainSemiProductPrice',
                    dataType: 'string',
                    isPrimitiveType: true,
                    isListContainer: false,
                    complexType: ''
                },
            ],
            validators: {
                semiProductId: [
                        ['required'],
                ],
                price: [
                        ['required'],
                ],
                currency: [
                        ['required'],
                ],
            }
        }
    }

  // export const ChainSemiProductPriceValidationScheme = {
  //     validators: [],
  //     fields: {
  //               semiProductId: {
  //                   validators: []
  //               },
  //               price: {
  //                   validators: []
  //               },
  //               currency: {
  //                   validators: []
  //               },
  //     }
  // } as SimpleValidationScheme<ChainSemiProductPrice>;


}